requirements.txt
djangorestframework==3.12.4
dj-database-url==0.5.0
Django==2.2.4
django-adminactions==1.6.0
envparse==0.2.0
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
unicodecsv==0.14.1
